Move mocked ftp to a sub-module
